fixed weight dtype conversion for counts in metrics

+++ makani/utils/metrics/functions.py
@@ -251,7 +251,7 @@ class GeometricSpread(GeometricBaseMetric):
             if self.batch_reduction == "mean":
                 raise ValueError(f"Batch reduction mode 'mean' is not supported when weights are provided. Use 'sum' instead.")
             elif self.batch_reduction == "sum":
-                counts = torch.sum(self.quadrature(weight), dim=0)
+                counts = torch.sum(self.quadrature(weight.to(dtype=inp.dtype)), dim=0)
             else:
                 raise ValueError(f"Batch reduction mode '{self.batch_reduction}' is not supported")
         else:
@@ -352,7 +352,7 @@ class GeometricSSR(GeometricBaseMetric):
             if self.batch_reduction == "mean":
                 raise ValueError(f"Batch reduction mode 'mean' is not supported when weights are provided. Use 'sum' instead.")
             elif self.batch_reduction == "sum":
-                counts = torch.sum(self.quadrature(weight), dim=0)
+                counts = torch.sum(self.quadrature(weight.to(dtype=inp.dtype)), dim=0)
             else:
                 raise ValueError(f"Batch reduction mode '{self.batch_reduction}' is not supported")
         else:
@@ -472,7 +472,7 @@ class GeometricCRPS(torch.nn.Module):
             if self.batch_reduction == "mean":
                 raise ValueError(f"Batch reduction mode 'mean' is not supported when weights are provided. Use 'sum' instead.")
             elif self.batch_reduction == "sum":
-                counts = torch.sum(self.quadrature(weight), dim=0)
+                counts = torch.sum(self.quadrature(weight.to(dtype=inp.dtype)), dim=0)
             else:
                 raise ValueError(f"Batch reduction mode '{self.batch_reduction}' is not supported")
         else:
@@ -558,7 +558,7 @@ class GeometricRankHistogram(GeometricBaseMetric):
             if self.batch_reduction == "mean":
                 raise ValueError(f"Batch reduction mode 'mean' is not supported when weights are provided. Use 'sum' instead.")
             elif self.batch_reduction == "sum":
-                counts = torch.sum(self.quadrature(weight), dim=0)
+                counts = torch.sum(self.quadrature(weight.to(dtype=inp.dtype)), dim=0)
             else:
                 raise ValueError(f"Batch reduction mode '{self.batch_reduction}' is not supported")
         else:
